requirements.txt
requests
qrcode[pil]
Pillow